% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nBootstrap-ci_bc.R
\name{ci_bc}
\alias{ci_bc}
\title{Bias-Corrected Confidence Intervals}
\usage{
ci_bc(x, thetahat, alpha = c(0.05, 0.01, 0.001))
}
\arguments{
\item{x}{Matrix, data frame, or vector.
Vector or matrix of estimates.}

\item{thetahat}{Numeric vector.
Vector of estimates from the original sample data.}

\item{alpha}{Numeric vector.
Significance level/s.}
}
\value{
A matrix.
}
\description{
Bias-Corrected Confidence Intervals
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\examples{
x <- rnorm(n = 100)
func <- median
thetahat <- func(x)
thetahatstar <- thetahatstar_nb(x, func = func)

ci_bc(
  thetahatstar,
  thetahat = thetahat
)
}
\seealso{
Other Nonparametric Bootstrap Functions: 
\code{\link{ci_bca_helper}()},
\code{\link{ci_bca}()},
\code{\link{ci_pc}()},
\code{\link{func_jackknife}()},
\code{\link{func_thetahatstar_nb}()},
\code{\link{func_xstar_nb}()},
\code{\link{jackknife}()},
\code{\link{thetahatstar_nb}()},
\code{\link{xstar_nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Nonparametric Bootstrap Functions}
\keyword{nBootstrap}
