% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nBootstrap-ci_pc.R
\name{ci_pc}
\alias{ci_pc}
\title{Percentile Confidence Intervals}
\usage{
ci_pc(x, alpha = c(0.05, 0.01, 0.001))
}
\arguments{
\item{x}{Matrix, data frame, or vector.
Vector or matrix of estimates.}

\item{alpha}{Numeric vector.
Significance level/s.}
}
\value{
A matrix.
}
\description{
Percentile Confidence Intervals
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\examples{
foo <- function(x) {
  median(rnorm(n = 100))
}
thetahatstar <- lapply(X = seq_len(1000), FUN = foo)
thetahatstar <- do.call(what = "rbind", args = thetahatstar)

ci_pc(thetahatstar)
}
\seealso{
Other Nonparametric Bootstrap Functions: 
\code{\link{ci_bca_helper}()},
\code{\link{ci_bca}()},
\code{\link{ci_bc}()},
\code{\link{func_jackknife}()},
\code{\link{func_thetahatstar_nb}()},
\code{\link{func_xstar_nb}()},
\code{\link{jackknife}()},
\code{\link{thetahatstar_nb}()},
\code{\link{xstar_nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Nonparametric Bootstrap Functions}
\keyword{nBootstrap}
