% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nBootstrap-func_xstar_nb.R
\name{func_xstar_nb}
\alias{func_xstar_nb}
\title{Constructor for the xstar_nb Function}
\usage{
func_xstar_nb(x)
}
\arguments{
\item{x}{Matrix, data frame, or vector.
Sample data.}
}
\value{
A function with \code{x} in the environment.
}
\description{
Constructor for the xstar_nb Function
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\examples{
x <- rnorm(n = 100)

func_xstar_nb(x)
}
\seealso{
Other Nonparametric Bootstrap Functions: 
\code{\link{ci_bca_helper}()},
\code{\link{ci_bca}()},
\code{\link{ci_bc}()},
\code{\link{ci_pc}()},
\code{\link{func_jackknife}()},
\code{\link{func_thetahatstar_nb}()},
\code{\link{jackknife}()},
\code{\link{thetahatstar_nb}()},
\code{\link{xstar_nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Nonparametric Bootstrap Functions}
\keyword{nBootstrap}
