% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nBootstrap-jackknife.R
\name{jackknife}
\alias{jackknife}
\title{Jackknife}
\usage{
jackknife(x, func, ...)
}
\arguments{
\item{x}{Matrix, data frame, or vector.
Sample data.}

\item{func}{Fit function.
The first argument should be sample data.}

\item{...}{Additional named arguments to pass to \code{func}.}
}
\value{
Returns a matrix if \code{func} returns a vector.
Otherwise, it returns a list.
}
\description{
Jackknife
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\examples{
x <- rnorm(n = 100)
func <- median

output <- jackknife(x, func = func)
hist(output)
}
\seealso{
Other Nonparametric Bootstrap Functions: 
\code{\link{ci_bca_helper}()},
\code{\link{ci_bca}()},
\code{\link{ci_bc}()},
\code{\link{ci_pc}()},
\code{\link{func_jackknife}()},
\code{\link{func_thetahatstar_nb}()},
\code{\link{func_xstar_nb}()},
\code{\link{thetahatstar_nb}()},
\code{\link{xstar_nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Nonparametric Bootstrap Functions}
\keyword{nBootstrap}
