% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nBootstrap-package.R
\docType{package}
\name{nBootstrap}
\alias{nBootstrap}
\title{nBootstrap: Nonparametric Bootstrap}
\description{
A collection of functions
related to the nonparametric bootstrap.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{nBootstrap}
\keyword{package}
