% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nBootstrap-xstar_nb.R
\name{xstar_nb}
\alias{xstar_nb}
\title{Generate Nonparametric Bootstrap Samples}
\usage{
xstar_nb(x, bcap = 5000L, seed = NULL)
}
\arguments{
\item{x}{Matrix, data frame, or vector.
Sample data.}

\item{bcap}{Integer.
Number of bootstrap samples.}

\item{seed}{Integer.
Random seed.}
}
\value{
Returns the same data type as input data.
}
\description{
Generate Nonparametric Bootstrap Samples
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\examples{
x <- rnorm(n = 100)
xstar_nb(x, bcap = 5L)
}
\seealso{
Other Nonparametric Bootstrap Functions: 
\code{\link{ci_bca_helper}()},
\code{\link{ci_bca}()},
\code{\link{ci_bc}()},
\code{\link{ci_pc}()},
\code{\link{func_jackknife}()},
\code{\link{func_thetahatstar_nb}()},
\code{\link{func_xstar_nb}()},
\code{\link{jackknife}()},
\code{\link{thetahatstar_nb}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Nonparametric Bootstrap Functions}
\keyword{nBootstrap}
